requirements.txt
discord.py
requests
datetime
geopy